Implement email notification service

+++ routes/personal-class.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const client = require('../database/handler');
+const mailSender = require('../utils/email-sender');
 
 router.get('/', async (req, res) => {
     try {
@@ -19,6 +20,11 @@ router.put('/', async (req, res) => {
             'INSERT INTO personal_class(start, end_, title, color, id, cssclass) VALUES($1, $2, $3, $4, $5, $6) RETURNING *',
             [body.start, body.end, body.title, body.color, body.id, body.cssClass]
         );
+
+        const start = new Date(body.start).toLocaleString('HU-hu');
+        const end = new Date(body.end).toLocaleString('HU-hu');
+        mailSender.sendMail('<pre>' + body.title + ' ' + start + '-tól ' + end + '-ig</pre>', body.title, 'preszl.daniel@gmail.com');
+
         res.send({serial_id: result.rows[0].serial_id});
     } catch (e) {
         console.log(e.stack);
@@ -28,7 +34,17 @@ router.put('/', async (req, res) => {
 
 router.delete('/*', async (req, res) => {
     try {
-        await client.query('DELETE FROM personal_class WHERE serial_id=' + req.path.substring(1));
+        const result = await client.query('SELECT start, end_, title FROM personal_class WHERE serial_id=' + req.path.substring(1));
+
+        if (result) {
+            await client.query('DELETE FROM personal_class WHERE serial_id=' + req.path.substring(1));
+
+            const start = new Date(result.rows[0].start).toLocaleString('HU-hu');
+            const end = new Date(result.rows[0].end_).toLocaleString('HU-hu');
+            mailSender.sendMail('<pre>' + result.rows[0].title + ' ' + start + '-tól ' + end + '-ig <b>TÖRÖLVE</b> </pre>',
+                result.rows[0].title + ' - törölve', 'preszl.daniel@gmail.com');
+        }
+
         res.send({serial_id: req.path.substring(1)});
     } catch (e) {
         console.log(e.stack);
